refactor: enforce new eslint barrel rules
