+++ discojs/src/client/decentralized/index.ts
@@ -1,2 +1,3 @@
 export { DecentralizedClient } from "./decentralized_client.js";
+// eslint-disable-next-line no-restricted-syntax -- namespace re-export acceptable here
 export * as messages from "./messages.js";
