+++ discojs/src/client/index.ts
@@ -10,6 +10,8 @@ export {
   messages as federatedMessages,
   FederatedClient,
 } from "./federated/index.js";
+//
+// eslint-disable-next-line no-restricted-syntax -- namespace re-export acceptable here
 export * as mtype from "./mtype.js";
 export { getClient } from "./get_client.js";
