+++ discojs/src/client/local_client.ts
@@ -1,5 +1,5 @@
 import type { WeightsContainer } from "#weights/index";
-import { Client } from "./client.js";
+import { Client } from "#client/client";
 
 /**
  * A LocalClient represents a Disco user training only on their local data without collaborating
