+++ discojs/src/index.ts
@@ -72,5 +72,6 @@ export type { DataType, Network, DataFormat } from "./types/index.js";
 
 export { extractColumn } from "./processing/index.js";
 
+// eslint-disable-next-line no-restricted-syntax -- namespace re-export acceptable here
 export * as defaultTasks from "./default_tasks/index.js";
 export { cards as defaultModels } from "./models/index.js";
