+++ discojs/src/models/onnx.spec.ts
@@ -2,8 +2,8 @@ import { describe, expect, it } from "vitest";
 
 import { List } from "immutable";
 import { AutoTokenizer } from "@xenova/transformers";
-import { ONNXModel } from "./onnx.js";
-import { DefaultGenerationConfig } from "./implementations/gpt/config.js";
+import { ONNXModel } from "#models/onnx";
+import { DefaultGenerationConfig } from "#models/implementations/gpt/config";
 
 describe("ONNXModel.predict", { timeout: 50_000 }, () => {
   it("should generate the next token ID from a prompt", async () => {
