+++ discojs/src/models/tfjs.ts
@@ -1,13 +1,13 @@
 import { List, Map, Range } from "immutable";
 import * as tf from "@tensorflow/tfjs";
 
-import { WeightsContainer } from "#weights/index";
 import type { Dataset, Batched } from "#dataset/index";
 import type { DataFormat, DataType } from "#types/index";
+import { WeightsContainer } from "#weights/index";
 
-import { Model } from "./model.js";
-import type { BatchLogs } from "./logs.js";
-import { EpochLogs } from "./logs.js";
+import type { BatchLogs } from "#models/logs";
+import { Model } from "#models/model";
+import { EpochLogs } from "#models/logs";
 
 type Serialized<D extends DataType> = [D, tf.io.ModelArtifacts];
