+++ discojs/src/processing/image.spec.ts
@@ -1,7 +1,9 @@
 import { Repeat, Seq } from "immutable";
 import { describe, expect, it } from "vitest";
+
 import { Image } from "#dataset/index";
-import { removeAlpha, resize } from "./image.js";
+
+import { removeAlpha, resize } from "#processing/image";
 
 describe("resize", () => {
   it("doesn't change with same image dimensions", () => {
