+++ discojs/src/serialization/weights.ts
@@ -2,8 +2,11 @@ import * as tf from "@tensorflow/tfjs";
 
 import { WeightsContainer } from "#weights/index";
 
-import type { Encoded } from "./coder.js";
-import * as coder from "./coder.js";
+import type { Encoded } from "#serialization/coder";
+import {
+  encode as encodeGeneric,
+  decode as decodeGeneric,
+} from "#serialization/coder";
 
 type Serialized = {
   shape: number[];
@@ -34,11 +37,11 @@ export async function encode(weights: WeightsContainer): Promise<Encoded> {
     })),
   );
 
-  return coder.encode(serialized);
+  return encodeGeneric(serialized);
 }
 
 export function decode(encoded: Encoded): WeightsContainer {
-  const raw = coder.decode(encoded);
+  const raw = decodeGeneric(encoded);
 
   if (!(Array.isArray(raw) && raw.every(isSerialized)))
     throw new Error("expected to decode an array of serialized weights");
