+++ discojs/src/task/task.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 
 import type { DataType, Network } from "#types/index";
 
-import { DisplayInformation } from "./display_information.js";
-import { TrainingInformation } from "./training_information.js";
+import { DisplayInformation } from "#task/display_information";
+import { TrainingInformation } from "#task/training_information";
 
 export namespace Task {
   export type ID = string;
