+++ discojs/src/types/datatype.ts
@@ -1,0 +1,2 @@
+export const dataTypeValues = ["image", "tabular", "text"] as const;
+export type DataType = (typeof dataTypeValues)[number];
