+++ discojs/src/types/index.ts
@@ -1,5 +1,7 @@
+// eslint-disable-next-line no-restricted-syntax -- namespace re-export acceptable here
 export * as DataFormat from "./data_format.js";
 
-export const dataTypeValues = ["image", "tabular", "text"] as const;
-export type DataType = (typeof dataTypeValues)[number];
-export type Network = "decentralized" | "federated" | "local";
+export { dataTypeValues } from "./datatype.js";
+export type { DataType } from "./datatype.js";
+
+export type { Network } from "./network.js";
