+++ discojs/src/types/network.ts
@@ -1,0 +1,1 @@
+export type Network = "decentralized" | "federated" | "local";
